converted stat change to typescript
